Started adding multiple book types
Added item_type to book copy object. It can be "book", "ebook", or
"audiobook". I added a dropdown to the search for books "Add to Library"
button, which allows the user to select which type of book it is. There
is still a lot of code that needs to be updated, but it's a start and
works to the point of adding a book from the search page and viewing it
in the library.