Added Discover Filtering
Discover page now has browse side links for each type of book. When
clicked they filter the results to the specific type. Also, all books
now filters out duplicates (where the OLKey is the same, but the
item_type is different--pure duplicates should be removed from the array
in Python, but this code isn't implemented yet).