Updated File/Folder Naming to eliminate "book"
src\books is now src\items
searchbooks.html is now search.html

All the code changes necessary to make this happen are in the previous
commit. The name changes are in a separate commit so the code changes
can be seen within GitHub (renaming a file only shows deletion of the
old and creation of the new--no comparison between).