Fixed Add to Library Button Bug (Changed Sizes)
The Add to Library button would change sizes whenever a tooltip
appeared, so I placed the tooltip in the button group instead of on the
add to library button. This causes the tooltip to appear even when the
mouse is over the dropdown button (unfortunate), and required that I
move the tooltips to above and below the buttons (otherwise it would
cover the dropdown button).