Fixed Cache and Discover Bugs, Updated RT API Key
Cache Bug: Updating the cache previously created a new Item object in
the datastore, rather than updating the current object. Now if cache is
true, the Item object is queried and updated, or if it returns none, a
new object is created. Tested for books only, but implemented for
standard movie queries and RT_Key queries.

Discover Bug: librarylist was list instead of
dict(item_key,item_subtype), causing the check for items in the user's
library to always return false.

RT API Key: The rotten tomatoes API key has been replaced with a new
key.