Fixed default name and custom URL bugs
Google's provided nickname is sometimes a long URL, so if it starts with
"http", is longer than 25 characters, or includes an @ sign (likely the
email address itself), I know make the default user's name the first
part of their provided email (everything up until the @).

If a user doesn't set a custom profile URL, but makes other changes to
settings, the form now omits the custom url and their user object
continues to contain "None" for the custom_url field, rather than
storing their ID in the field.